LayoutTests/platform/wpe/imported/w3c/web-platform-tests/html/editing/dnd/datastore/datatransfer-constructor-001-expected.txt


FAIL Verify DataTransfer constructor assert_equals: expected "none" but got "uninitialized"
